+++ pykalman/utils/__init__.py
@@ -1,0 +1,21 @@
+"""Utilities for the package."""
+
+from pykalman.utils._sklearn import (
+    Bunch,
+    array1d,
+    array2d,
+    check_random_state,
+    get_params,
+    log_multivariate_normal_density,
+    preprocess_arguments,
+)
+
+__all__ = [
+    "Bunch",
+    "array1d",
+    "array2d",
+    "check_random_state",
+    "get_params",
+    "log_multivariate_normal_density",
+    "preprocess_arguments",
+]
